\name{columbus}
\title{Columbus, OH crime data}
\usage{data(columbus)}
\alias{columbus}
\format{A data frame with 49 observations on 5 variables.
  \tabular{rlll}{
    [,1]  \tab crime  \tab numeric  \tab recorded crime per inhabitant\cr
    [,2]  \tab income  \tab numeric  \tab average income values\cr
    [,3]  \tab housing  \tab numeric  \tab average housing costs\cr
    [,4]  \tab x  \tab numeric  \tab Easting\cr
    [,5]  \tab y   \tab numeric  \tab Northing\cr
  }
}
\source{Luc Anselin (1995): SpaceStat
}
\description{Columbus, OH crime data

}

\note{There are two extant versions of these data - these give the exact results reproduced in the SpaceStat manual}
\references{Luc Anselin (1995): SpaceStat}

\keyword{datasets}
